Moved the piano classes to their own package
