+++ .idea/workspace.xml
@@ -50,7 +50,7 @@
       <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
       <extension name="coverage">
         <pattern>
-          <option name="PATTERN" value="app.ui.*" />
+          <option name="PATTERN" value="app.ui.piano.*" />
           <option name="ENABLED" value="true" />
         </pattern>
       </extension>
@@ -121,21 +121,21 @@
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
     <state x="740" y="275" key="FileChooserDialogImpl/0.0.1920.1040@0.0.1920.1040" timestamp="1588689813658" />
-    <state width="1877" height="281" key="GridCell.Tab.0.bottom" timestamp="1588690265852">
+    <state width="1877" height="281" key="GridCell.Tab.0.bottom" timestamp="1588690406088">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1877" height="281" key="GridCell.Tab.0.bottom/0.0.1920.1040@0.0.1920.1040" timestamp="1588690265852" />
-    <state width="1877" height="281" key="GridCell.Tab.0.center" timestamp="1588690265852">
+    <state width="1877" height="281" key="GridCell.Tab.0.bottom/0.0.1920.1040@0.0.1920.1040" timestamp="1588690406088" />
+    <state width="1877" height="281" key="GridCell.Tab.0.center" timestamp="1588690406088">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1877" height="281" key="GridCell.Tab.0.center/0.0.1920.1040@0.0.1920.1040" timestamp="1588690265852" />
-    <state width="1877" height="281" key="GridCell.Tab.0.left" timestamp="1588690265851">
+    <state width="1877" height="281" key="GridCell.Tab.0.center/0.0.1920.1040@0.0.1920.1040" timestamp="1588690406088" />
+    <state width="1877" height="281" key="GridCell.Tab.0.left" timestamp="1588690406088">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1877" height="281" key="GridCell.Tab.0.left/0.0.1920.1040@0.0.1920.1040" timestamp="1588690265851" />
-    <state width="1877" height="281" key="GridCell.Tab.0.right" timestamp="1588690265852">
+    <state width="1877" height="281" key="GridCell.Tab.0.left/0.0.1920.1040@0.0.1920.1040" timestamp="1588690406088" />
+    <state width="1877" height="281" key="GridCell.Tab.0.right" timestamp="1588690406088">
       <screen x="0" y="0" width="1920" height="1040" />
     </state>
-    <state width="1877" height="281" key="GridCell.Tab.0.right/0.0.1920.1040@0.0.1920.1040" timestamp="1588690265852" />
+    <state width="1877" height="281" key="GridCell.Tab.0.right/0.0.1920.1040@0.0.1920.1040" timestamp="1588690406088" />
   </component>
 </project>
